Move the platform value into a platform module
