+++ packages/react-native-reanimated/src/createAnimatedComponent/platform.native.ts
