+++ packages/react-native-reanimated/src/createAnimatedComponent/platform.ts
